UserController third commit delete package mappers, add UserDtoDao UserDtoService

+++ src/main/java/com/javamentor/qa/platform/models/dto/UserDto.java
@@ -13,5 +13,5 @@ public class UserDto implements Serializable {
     private String email;
     private String fullName;
     private String linkImage;
-    private Long reputation;
+    private int reputation;
 }
